+++ test/hexapdf/type/test_annotation.rb
@@ -4,6 +4,29 @@ require 'test_helper'
 require 'hexapdf/document'
 require 'hexapdf/type/annotation'
 
+describe HexaPDF::Type::Annotation::AppearanceDictionary do
+  before do
+    @doc = HexaPDF::Document.new
+    @ap = @doc.add({N: :n, D: :d, R: :r}, type: :XXAppearanceDictionary)
+  end
+
+  it "resolves the normal appearance" do
+    assert_equal(:n, @ap.normal_appearance)
+  end
+
+  it "resolves the rollover appearance" do
+    assert_equal(:r, @ap.rollover_appearance)
+    @ap.delete(:R)
+    assert_equal(:n, @ap.rollover_appearance)
+  end
+
+  it "resolves the down appearance" do
+    assert_equal(:d, @ap.down_appearance)
+    @ap.delete(:D)
+    assert_equal(:n, @ap.down_appearance)
+  end
+end
+
 describe HexaPDF::Type::Annotation do
   before do
     @doc = HexaPDF::Document.new