+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   [HexaPDF::Type::AcroForm::Field]
 * [HexaPDF::Type::AcroForm::TextField] and
   [HexaPDF::Type::AcroForm::VariableTextField] for basic text field support
+* [HexaPDF::Type::Annotation::AppearanceDictionary]
 
 
 ## 0.11.9 - 2020-06-15